term additio for hard deletion & small variable name changes
